refactor node templates for code reuse
